refactor(assistant): split provider config tests
- move assistant provider environment selection coverage into a provider-local test file
- leave handler tests focused on request handling and streaming behavior
- reduce the handler test hotspot score from 34.8 to 30.4 while preserving coverage
